pref(ripple): 文件格式
