--- packages/devui-vue/devui/ripple/src/utils/getrelative-pointer.ts
@@ -1,7 +1,0 @@
-export const getRelativePointer = (
-  { x, y }: PointerEvent,
-  { top, left }: DOMRect
-): Partial<DOMRect> => ({
-  x: x - left,
-  y: y - top
-});
